address review and add one more test
