+++ packages/yew/src/app_handle.rs
@@ -90,7 +90,7 @@ mod feat_hydration {
                 NodeRef::default(),
                 Rc::clone(&props),
             );
-            // Fix next_sibling at the root
+            #[cfg(debug_assertions)] // Fix trapped next_sibling at the root
             app.scope.reuse(props, NodeRef::default());
 
             // We remove all remaining nodes, this mimics the clear_element behaviour in
